% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_cascade_convergence}
\alias{ccf_cascade_convergence}
\title{Cascade convergence statistics for a year.}
\usage{
ccf_cascade_convergence(client, year)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{year}{Integer.}
}
\description{
Cascade convergence statistics for a year.
}
