% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_cascade_detail}
\alias{ccf_cascade_detail}
\title{Detail for a single cascade.}
\usage{
ccf_cascade_detail(client, year, cascade_id)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{year}{Integer.}

\item{cascade_id}{Cascade identifier.}
}
\description{
Detail for a single cascade.
}
