% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_cascade_events}
\alias{ccf_cascade_events}
\title{Event clusters listed under a year's cascade data.}
\usage{
ccf_cascade_events(client, year, raw = FALSE)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{year}{Integer.}

\item{raw}{If \code{TRUE} return raw list.}
}
\description{
Event clusters listed under a year's cascade data.
}
