% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_cascade_impact}
\alias{ccf_cascade_impact}
\title{Cascade impact summary for a year.}
\usage{
ccf_cascade_impact(client, year)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{year}{Integer.}
}
\description{
Cascade impact summary for a year.
}
