% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_cascade_network}
\alias{ccf_cascade_network}
\title{Network edges for a single cascade.}
\usage{
ccf_cascade_network(client, year, cascade_id)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{year}{Integer.}

\item{cascade_id}{Cascade ID.}
}
\description{
Network edges for a single cascade.
}
