% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_cascade_paradigm_shifts}
\alias{ccf_cascade_paradigm_shifts}
\title{Paradigm-shift episodes for a year.}
\usage{
ccf_cascade_paradigm_shifts(client, year)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{year}{Integer.}
}
\description{
Paradigm-shift episodes for a year.
}
