% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_cascade_time_series}
\alias{ccf_cascade_time_series}
\title{Daily time-series tables for a year.}
\usage{
ccf_cascade_time_series(client, year)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{year}{Integer.}
}
\description{
Daily time-series tables for a year.
}
