% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_cascade_year}
\alias{ccf_cascade_year}
\title{All cascades for a given year.}
\usage{
ccf_cascade_year(client, year)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{year}{Integer year.}
}
\description{
All cascades for a given year.
}
