% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_cascade_year_network}
\alias{ccf_cascade_year_network}
\title{Whole-year network with optional filters.}
\usage{
ccf_cascade_year_network(
  client,
  year,
  cascade_ids = NULL,
  frames = NULL,
  media = NULL,
  classifications = NULL,
  score_min = NULL,
  score_max = NULL
)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{year}{Integer.}

\item{cascade_ids, frames, media, classifications}{Character vectors of filters (optional).}

\item{score_min, score_max}{Numeric (optional).}
}
\description{
Whole-year network with optional filters.
}
