% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_cascades_cross_year}
\alias{ccf_cascades_cross_year}
\title{Cross-year cascade table (paginated).}
\usage{
ccf_cascades_cross_year(client, page = 1L, page_size = 100L, raw = FALSE)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{page, page_size}{Pagination.}

\item{raw}{If \code{TRUE} return raw response.}
}
\description{
Cross-year cascade table (paginated).
}
