% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_cascades_cross_year_all}
\alias{ccf_cascades_cross_year_all}
\title{All cross-year cascades (slim metadata, single response).}
\usage{
ccf_cascades_cross_year_all(client, raw = FALSE)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{raw}{If \code{TRUE} return raw response.}
}
\description{
All cross-year cascades (slim metadata, single response).
}
