% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_cascades_paradigm_timeline}
\alias{ccf_cascades_paradigm_timeline}
\title{Cross-year paradigm-shift timeline.}
\usage{
ccf_cascades_paradigm_timeline(client, raw = FALSE)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{raw}{If \code{TRUE} return raw response.}
}
\description{
Cross-year paradigm-shift timeline.
}
