% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_cascades_search}
\alias{ccf_cascades_search}
\title{Keyword/metadata search over cascades.}
\usage{
ccf_cascades_search(
  client,
  query,
  mode = "text",
  cascade_id = NULL,
  raw = FALSE
)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{query}{Search text.}

\item{mode}{\code{"text"} or \code{"similar"} (with \code{cascade_id}).}

\item{cascade_id}{Used when \code{mode = "similar"}.}

\item{raw}{If \code{TRUE} return raw response.}
}
\description{
Keyword/metadata search over cascades.
}
