% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_cascades_semantic_search}
\alias{ccf_cascades_semantic_search}
\title{FAISS-backed semantic search for cascades.}
\usage{
ccf_cascades_semantic_search(client, query, k = 100000L, raw = FALSE)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{query}{Search text.}

\item{k}{Top-k neighbours.}

\item{raw}{If \code{TRUE} return raw response.}
}
\description{
FAISS-backed semantic search for cascades.
}
