% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_cascades_summary}
\alias{ccf_cascades_summary}
\title{Cross-year cascade summary.}
\usage{
ccf_cascades_summary(client)
}
\arguments{
\item{client}{A \code{ccf_client}.}
}
\description{
Cross-year cascade summary.
}
