% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{ccf_codebook_df}
\alias{ccf_codebook_df}
\title{Codebook as a tidy tibble.}
\usage{
ccf_codebook_df()
}
\value{
A tibble with columns \code{column}, \code{group}, \code{subgroup}, \code{definition}.
}
\description{
Codebook as a tidy tibble.
}
