% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_cross_tabulation}
\alias{ccf_cross_tabulation}
\title{2x2 contingency table of two binary annotation columns.}
\usage{
ccf_cross_tabulation(client, row_var, col_var, filters = list(), corpus = NULL)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{row_var, col_var}{Column names.}

\item{filters}{Optional named list of filters.}
}
\description{
2x2 contingency table of two binary annotation columns.
}
