% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{ccf_define}
\alias{ccf_define}
\title{Operational definition for a single annotation column.}
\usage{
ccf_define(column)
}
\arguments{
\item{column}{Annotation column name (e.g. \code{"eco_neg_impact"}).}
}
\value{
Character scalar.
}
\description{
Operational definition for a single annotation column.
}
\examples{
ccf_define("sci_skepticism")
}
