% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_events_cluster}
\alias{ccf_events_cluster}
\title{Single event-cluster detail.}
\usage{
ccf_events_cluster(client, year, cluster_id)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{year, cluster_id}{Integer.}
}
\description{
Single event-cluster detail.
}
