% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_events_cluster_articles}
\alias{ccf_events_cluster_articles}
\title{Articles attached to an event cluster (grouped by occurrence).}
\usage{
ccf_events_cluster_articles(client, year, cluster_id)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{year, cluster_id}{Integer.}
}
\description{
Articles attached to an event cluster (grouped by occurrence).
}
