% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_events_search}
\alias{ccf_events_search}
\title{Keyword search for event clusters.}
\usage{
ccf_events_search(
  client,
  query,
  types = NULL,
  year_min = NULL,
  year_max = NULL,
  strength_min = NULL,
  raw = FALSE
)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{query}{Search text.}

\item{types}{Character vector of event types.}

\item{year_min, year_max, strength_min}{Optional filters.}

\item{raw}{If \code{TRUE} return raw response.}
}
\description{
Keyword search for event clusters.
}
