% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_events_summary}
\alias{ccf_events_summary}
\title{Cross-year event-cluster summary.}
\usage{
ccf_events_summary(client)
}
\arguments{
\item{client}{A \code{ccf_client}.}
}
\description{
Cross-year event-cluster summary.
}
