% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_events_type_network}
\alias{ccf_events_type_network}
\title{Co-occurrence matrix between event types.}
\usage{
ccf_events_type_network(client, year = NULL)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{year}{Optional integer to filter.}
}
\description{
Co-occurrence matrix between event types.
}
