% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{ccf_frame_names}
\alias{ccf_frame_names}
\alias{ccf_frame_columns}
\title{Frame metadata.}
\usage{
ccf_frame_names()

ccf_frame_columns()
}
\value{
Character vector of frame names (length 8).

Character vector of frame DB columns (length 8).
}
\description{
Frame metadata.
}
