% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_geo_data}
\alias{ccf_geo_data}
\title{Server-side annotation schema (frames, subcategories, columns).}
\usage{
ccf_geo_data(client)
}
\arguments{
\item{client}{A \code{ccf_client}.}
}
\description{
Server-side annotation schema (frames, subcategories, columns).
}
