% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_annotation_metrics}
\alias{ccf_live_annotation_metrics}
\title{Per-category annotation performance metrics of the CCF models.}
\usage{
ccf_live_annotation_metrics(client)
}
\arguments{
\item{client}{A \code{ccf_live} client.}
}
\description{
Per-category annotation performance metrics of the CCF models.
}
