% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_articles_by_media}
\alias{ccf_live_articles_by_media}
\title{Article counts per outlet (live corpus).}
\usage{
ccf_live_articles_by_media(client)
}
\arguments{
\item{client}{A \code{ccf_live} client.}
}
\description{
Article counts per outlet (live corpus).
}
