% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_articles_by_month}
\alias{ccf_live_articles_by_month}
\title{Article counts per month (live corpus).}
\usage{
ccf_live_articles_by_month(client)
}
\arguments{
\item{client}{A \code{ccf_live} client.}
}
\description{
Article counts per month (live corpus).
}
