% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_cascade}
\alias{ccf_live_cascade}
\title{Full profile of one cascade (articles, outlets, summaries EN/FR).}
\usage{
ccf_live_cascade(client, cascade_id)
}
\arguments{
\item{client}{A \code{ccf_live} client.}

\item{cascade_id}{Cascade identifier.}
}
\description{
Full profile of one cascade (articles, outlets, summaries EN/FR).
}
