% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_cascade_summary}
\alias{ccf_live_cascade_summary}
\title{Aggregate cascade statistics for the observatory.}
\usage{
ccf_live_cascade_summary(client)
}
\arguments{
\item{client}{A \code{ccf_live} client.}
}
\description{
Aggregate cascade statistics for the observatory.
}
