% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_category_distribution}
\alias{ccf_live_category_distribution}
\title{Distribution of detected categories across the corpus.}
\usage{
ccf_live_category_distribution(client)
}
\arguments{
\item{client}{A \code{ccf_live} client.}
}
\description{
Distribution of detected categories across the corpus.
}
