% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_frames_by_media}
\alias{ccf_live_frames_by_media}
\title{Monthly frame shares aggregated by outlet.}
\usage{
ccf_live_frames_by_media(client)
}
\arguments{
\item{client}{A \code{ccf_live} client.}
}
\description{
Monthly frame shares aggregated by outlet.
}
