% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_frames_data}
\alias{ccf_live_frames_data}
\title{Frame trend data used by the observatory charts.}
\usage{
ccf_live_frames_data(client)
}
\arguments{
\item{client}{A \code{ccf_live} client.}
}
\description{
Frame trend data used by the observatory charts.
}
