% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_frames_national}
\alias{ccf_live_frames_national}
\title{National monthly frame shares (smoothed), live corpus.}
\usage{
ccf_live_frames_national(client)
}
\arguments{
\item{client}{A \code{ccf_live} client.}
}
\description{
National monthly frame shares (smoothed), live corpus.
}
