% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_latest_articles}
\alias{ccf_live_latest_articles}
\title{Most recent extracted articles with their LLM summaries.}
\usage{
ccf_live_latest_articles(client)
}
\arguments{
\item{client}{A \code{ccf_live} client.}
}
\description{
Most recent extracted articles with their LLM summaries.
}
