% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_latest_classified}
\alias{ccf_live_latest_classified}
\title{Most recent articles fully classified by the 128 CCF models.}
\usage{
ccf_live_latest_classified(client)
}
\arguments{
\item{client}{A \code{ccf_live} client.}
}
\description{
Most recent articles fully classified by the 128 CCF models.
}
