% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_media_coverage}
\alias{ccf_live_media_coverage}
\title{Coverage freshness per outlet and per source database.}
\usage{
ccf_live_media_coverage(client)
}
\arguments{
\item{client}{A \code{ccf_live} client.}
}
\description{
Coverage freshness per outlet and per source database.
}
