% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_network_data}
\alias{ccf_live_network_data}
\title{Entity co-occurrence network of the live corpus.}
\usage{
ccf_live_network_data(client)
}
\arguments{
\item{client}{A \code{ccf_live} client.}
}
\description{
Entity co-occurrence network of the live corpus.
}
