% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_observatory_summary}
\alias{ccf_live_observatory_summary}
\title{Observatory headline summary.}
\usage{
ccf_live_observatory_summary(client)
}
\arguments{
\item{client}{A \code{ccf_live} client.}
}
\description{
Observatory headline summary.
}
