% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_ongoing_events}
\alias{ccf_live_ongoing_events}
\title{Events detected today or yesterday — the freshest signals.}
\usage{
ccf_live_ongoing_events(client)
}
\arguments{
\item{client}{A \code{ccf_live} client.}
}
\description{
Events detected today or yesterday — the freshest signals.
}
