% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_overview_summary}
\alias{ccf_live_overview_summary}
\title{LLM overview of the 20 biggest events of the last 15 days (EN/FR).}
\usage{
ccf_live_overview_summary(client)
}
\arguments{
\item{client}{A \code{ccf_live} client.}
}
\description{
LLM overview of the 20 biggest events of the last 15 days (EN/FR).
}
