% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_search_cascades}
\alias{ccf_live_search_cascades}
\title{Full-text search over cascade titles and summaries (EN + FR).}
\usage{
ccf_live_search_cascades(client, q)
}
\arguments{
\item{client}{A \code{ccf_live} client.}

\item{q}{Query string.}
}
\description{
Full-text search over cascade titles and summaries (EN + FR).
}
