% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_search_titles}
\alias{ccf_live_search_titles}
\title{Full-text search over article titles of the live corpus.}
\usage{
ccf_live_search_titles(client, q)
}
\arguments{
\item{client}{A \code{ccf_live} client.}

\item{q}{Query string.}
}
\description{
Full-text search over article titles of the live corpus.
}
