% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_stats}
\alias{ccf_live_stats}
\title{Site-wide corpus statistics (articles, sentences, time span).}
\usage{
ccf_live_stats(client)
}
\arguments{
\item{client}{A \code{ccf_live} client.}
}
\description{
Site-wide corpus statistics (articles, sentences, time span).
}
