% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{ccf_live_tone_over_time}
\alias{ccf_live_tone_over_time}
\title{Monthly tone (alarmist / reassuring) trends.}
\usage{
ccf_live_tone_over_time(client)
}
\arguments{
\item{client}{A \code{ccf_live} client.}
}
\description{
Monthly tone (alarmist / reassuring) trends.
}
