% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{ccf_media_outlets}
\alias{ccf_media_outlets}
\title{List of media outlets covered by the corpus.}
\usage{
ccf_media_outlets()
}
\value{
Character vector of length 20.
}
\description{
List of media outlets covered by the corpus.
}
