% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_search_summary}
\alias{ccf_search_summary}
\title{Aggregate stats for a search query (year/media distribution, frame breakdown).}
\usage{
ccf_search_summary(client, query, filters = list(), corpus = NULL)
}
\arguments{
\item{client}{A \code{ccf_client}.}

\item{query}{Search text.}

\item{filters}{Named list.}
}
\description{
Aggregate stats for a search query (year/media distribution, frame breakdown).
}
