% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{ccf_subcategories_of}
\alias{ccf_subcategories_of}
\title{Subcategory column names for a given frame.}
\usage{
ccf_subcategories_of(frame)
}
\arguments{
\item{frame}{One of \code{"economic"}, \code{"health"}, \code{"security"}, \code{"justice"},
\code{"political"}, \code{"scientific"}, \code{"environmental"}, \code{"cultural"}.}
}
\value{
Character vector of column names.
}
\description{
Subcategory column names for a given frame.
}
\examples{
ccf_subcategories_of("economic")
}
