% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiers.R
\name{ccf_tier_descriptions}
\alias{ccf_tier_descriptions}
\title{Short description of each tier.}
\usage{
ccf_tier_descriptions()
}
\value{
Named character vector indexed by tier name.
}
\description{
Short description of each tier.
}
