% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiers.R
\name{ccf_tier_names}
\alias{ccf_tier_names}
\title{Names of all CCF API tiers, in order of increasing privilege.}
\format{
Character vector of length 5.
}
\usage{
ccf_tier_names()
}
\description{
Names of all CCF API tiers, in order of increasing privilege.
}
