% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ccf_tiers}
\alias{ccf_tiers}
\title{Public listing of all tiers + their default quotas.}
\usage{
ccf_tiers(client)
}
\arguments{
\item{client}{A \code{ccf_client}.}
}
\description{
Public listing of all tiers + their default quotas.
}
